refactor(croquis): split script_parser into cohesive modules
Move the result/option types, global-scope setup, and public parse
entry points out of the script_parser module root into dedicated
submodules under script_parser/:

- result.rs: ScriptParseResult, ScriptParserOptions, and the
  ReactiveValueOrigin / ReactiveGetterContext / RuntimeObjectLiteral
  metadata types plus the impl ScriptParseResult block.
- globals.rs: browser/Node/Vue global name tables, build_global_names,
  and setup_global_scopes.
- parse.rs: parse_script_setup{,_with_generic}, parse_script,
  parse_script_with_options.

The module root keeps the docs, mod declarations, re-exports, and the
test module. All items are re-exported so crate::script_parser::* and
vize_croquis::script_parser::* paths resolve unchanged. Pure structural
move; no behavior, signature, or public API changes.
