+++ crates/vize_croquis/src/script_parser.rs
@@ -10,502 +10,27 @@
 //!
 //! ## Module Structure
 //!
+//! - [`parse`] - Public parse entry points (script setup / plain)
+//! - [`result`] - [`ScriptParseResult`] and option/metadata types
+//! - [`globals`] - Global-name tables and scope hierarchy setup
 //! - [`process`] - Statement and variable processing
 //! - [`extract`] - Props/emits extraction and reactivity detection
 //! - [`walk`] - Scope walking functions
 
 mod extract;
+mod globals;
+mod parse;
 mod process;
+mod result;
 mod typeof_refs;
 mod walk;
 
-use std::sync::LazyLock;
-
-use oxc_allocator::Allocator;
-use oxc_parser::Parser;
-use oxc_span::SourceType;
-
-use crate::croquis::{BindingMetadata, ComponentRegistration, Croquis};
-use crate::croquis::{ImportStatementInfo, InvalidExport, ReExportInfo, TypeExport};
-use crate::macros::{EmitDefinition, MacroTracker, PropDefinition};
-use crate::provide::ProvideInjectTracker;
-use crate::race::RaceConditionTracker;
-use crate::reactivity::ReactivityTracker;
-use crate::scope::{
-    JsGlobalScopeData, JsRuntime, NonScriptSetupScopeData, ParamNames, ScopeChain,
-    ScriptSetupScopeData, VueGlobalScopeData,
+pub use parse::{
+    parse_script, parse_script_setup, parse_script_setup_with_generic, parse_script_with_options,
 };
-use crate::setup_context::SetupContextTracker;
-use vize_carton::{CompactString, FxHashMap, FxHashSet, profile};
-
 pub use process::process_statement;
-
-/// Origin of a local binding that already carries a plain, non-reactive value.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum ReactiveValueOrigin {
-    PropsDestructure {
-        prop_name: CompactString,
-    },
-    ReactiveProperty {
-        source_name: CompactString,
-        prop_name: CompactString,
-    },
-    RefValue {
-        source_name: CompactString,
-    },
-    FunctionArgument {
-        source_name: CompactString,
-        callee_name: CompactString,
-    },
-    GetterCall {
-        context_name: CompactString,
-        getter_name: CompactString,
-        source_name: CompactString,
-    },
-    PlainAlias {
-        source_name: CompactString,
-    },
-}
-
-/// A returned context whose methods are backed by getter arguments.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct ReactiveGetterContext {
-    pub callee_name: CompactString,
-    pub getters: FxHashMap<CompactString, CompactString>,
-}
-
-/// Static metadata extracted from a top-level runtime object literal.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct RuntimeObjectLiteral {
-    pub props: Vec<PropDefinition>,
-    pub emits: Vec<EmitDefinition>,
-}
-
-/// Result of parsing a script setup block
-#[derive(Debug, Default)]
-pub struct ScriptParseResult {
-    pub bindings: BindingMetadata,
-    pub macros: MacroTracker,
-    pub reactivity: ReactivityTracker,
-    pub race_conditions: RaceConditionTracker,
-    pub type_exports: Vec<TypeExport>,
-    pub invalid_exports: Vec<InvalidExport>,
-    /// Scope chain for tracking nested JavaScript scopes
-    pub scopes: ScopeChain,
-    /// Provide/Inject tracking
-    pub provide_inject: ProvideInjectTracker,
-    /// Track inject variable names for indirect destructure detection
-    pub(crate) inject_var_names: FxHashSet<CompactString>,
-    /// Track aliases for inject function (e.g., const a = inject; a('key'))
-    pub(crate) inject_aliases: FxHashSet<CompactString>,
-    /// Track aliases for provide function (e.g., const p = provide; p('key', val))
-    pub(crate) provide_aliases: FxHashSet<CompactString>,
-    /// Track aliases for reactivity APIs (e.g., const r = ref; r(0))
-    /// Maps alias name to the original function name
-    pub(crate) reactivity_aliases: FxHashMap<CompactString, CompactString>,
-    /// Bindings that are known plain snapshots of reactive values.
-    pub(crate) reactive_value_origins: FxHashMap<CompactString, ReactiveValueOrigin>,
-    /// Call results that were constructed from getter arguments.
-    pub(crate) reactive_getter_contexts: FxHashMap<CompactString, ReactiveGetterContext>,
-    /// Setup context violation tracking
-    pub setup_context: SetupContextTracker,
-    /// Flag to track if we're in a non-setup script context
-    pub(crate) is_non_setup_script: bool,
-    /// Import statement spans in script content
-    pub import_statements: Vec<ImportStatementInfo>,
-    /// Re-export statement spans (`export { ... } from "..."`)
-    pub re_exports: Vec<ReExportInfo>,
-    /// Components registered through Options API `components`.
-    pub component_registrations: Vec<ComponentRegistration>,
-    /// Definition spans for bindings (name -> (start, end) offset in script)
-    pub binding_spans: FxHashMap<CompactString, (u32, u32)>,
-    /// Value import source by local binding name.
-    pub import_sources: FxHashMap<CompactString, CompactString>,
-    /// Names referenced via `typeof X` in the body of each `type_exports`
-    /// entry, indexed in parallel with `type_exports`. Used by
-    /// `resolve_type_export_hoisting` to keep types adjacent to the
-    /// setup-scope values they depend on. Pushed to in lockstep with
-    /// `type_exports` via `record_type_export`.
-    pub(crate) type_export_typeof_refs: Vec<FxHashSet<CompactString>>,
-    /// Static runtime object literal metadata available to macro spread args.
-    pub(crate) runtime_object_literals: FxHashMap<CompactString, RuntimeObjectLiteral>,
-}
-
-/// Options for plain script parsing.
-#[derive(Debug, Clone, Copy, Default)]
-pub struct ScriptParserOptions {
-    /// Resolve Options API template bindings (`data`/`computed`/`methods`/
-    /// `inject`/`setup`/`props`). This is officially supported in Vue 3 and is
-    /// an opt-in for the standard build — it does **not** require the `legacy`
-    /// feature.
-    pub options_api: bool,
-    /// Additionally treat the component as legacy Vue 2.7 / Nuxt 2: implies
-    /// `options_api` binding resolution and adds Nuxt 2 template globals.
-    pub legacy_vue2: bool,
-}
-
-impl ScriptParseResult {
-    /// Record a `TypeExport` together with the `typeof` value-identifier
-    /// references found in its body. Must be the only call site that pushes
-    /// to `type_exports` so the two vectors stay in lockstep for
-    /// `resolve_type_export_hoisting`.
-    pub(crate) fn record_type_export(
-        &mut self,
-        export: TypeExport,
-        typeof_refs: FxHashSet<CompactString>,
-    ) {
-        self.type_exports.push(export);
-        self.type_export_typeof_refs.push(typeof_refs);
-    }
-
-    /// Demote `TypeExport::hoisted` to `false` for any type whose body
-    /// references a setup-scope value binding via `typeof`. The virtual TS
-    /// generator only lifts hoisted types to module scope, so demoted types
-    /// stay inside the synthetic `__setup` function alongside the values
-    /// they depend on — which is the only place TS can resolve them.
-    ///
-    /// Imports add value bindings at module scope, so a `typeof importedName`
-    /// reference is left as hoisted: the import is visible from the module
-    /// scope where the type lands.
-    pub(crate) fn resolve_type_export_hoisting(&mut self) {
-        if self.type_export_typeof_refs.is_empty() {
-            return;
-        }
-
-        // A `typeof name` ref keeps a type hoisted only when `name` is a
-        // module-scoped import. Rather than materialize the full set of
-        // imported binding names up front (O(imports × bindings)), test each
-        // referenced name's declaration span against the import ranges on
-        // demand — `refs` is tiny, so this is O(refs × imports).
-        for idx in 0..self.type_export_typeof_refs.len() {
-            let touches_setup_value = self.type_export_typeof_refs[idx].iter().any(|name| {
-                let key = name.as_str();
-                self.bindings.bindings.contains_key(key)
-                    && !self.binding_spans.get(key).is_some_and(|(start, end)| {
-                        // Bindings whose declaration site falls inside an import
-                        // statement are module-scoped imports, not setup values.
-                        self.import_statements
-                            .iter()
-                            .any(|imp| *start >= imp.start && *end <= imp.end)
-                    })
-            });
-            if touches_setup_value && let Some(te) = self.type_exports.get_mut(idx) {
-                te.hoisted = false;
-            }
-        }
-    }
-
-    /// Apply script analysis fields to an existing SFC analysis summary.
-    ///
-    /// This keeps script parsing as the single owner of script-scoped data while
-    /// allowing callers to add template analysis before or after the script pass.
-    pub fn apply_to_croquis(self, summary: &mut Croquis) {
-        summary.bindings = self.bindings;
-        summary.macros = self.macros;
-        summary.reactivity = self.reactivity;
-        summary.race_conditions = self.race_conditions;
-        summary.type_exports = self.type_exports;
-        summary.invalid_exports = self.invalid_exports;
-        summary.scopes = self.scopes;
-        summary.provide_inject = self.provide_inject;
-        summary.setup_context = self.setup_context;
-        summary.import_statements = self.import_statements;
-        summary.re_exports = self.re_exports;
-        summary.component_registrations = self.component_registrations;
-        summary.binding_spans = self.binding_spans;
-    }
-
-    /// Convert script analysis into a `Croquis` summary.
-    pub fn into_croquis(self) -> Croquis {
-        let mut summary = Croquis::new();
-        self.apply_to_croquis(&mut summary);
-        summary
-    }
-}
-
-/// Browser-only globals (WHATWG Living Standard + HTML timers).
-///
-/// Immutable, identical for every file. Kept as a static `&[&str]` so the
-/// `CompactString` list (see `BROWSER_GLOBAL_NAMES`) is materialized once
-/// instead of re-running the `smallvec!`/`const_new` construction per file.
-static BROWSER_GLOBALS: &[&str] = &[
-    "alert",
-    "Audio",
-    "cancelAnimationFrame",
-    "cancelIdleCallback",
-    "CanvasRenderingContext2D",
-    "clearInterval",
-    "clearTimeout",
-    "close",
-    "confirm",
-    "customElements",
-    "document",
-    "Document",
-    "DocumentFragment",
-    "Element",
-    "FocusEvent",
-    "getComputedStyle",
-    "getSelection",
-    "history",
-    "HTMLElement",
-    "Image",
-    "indexedDB",
-    "InputEvent",
-    "IntersectionObserver",
-    "KeyboardEvent",
-    "localStorage",
-    "location",
-    "matchMedia",
-    "MediaQueryList",
-    "MouseEvent",
-    "MutationObserver",
-    "navigator",
-    "Node",
-    "NodeList",
-    "open",
-    "PerformanceObserver",
-    "PointerEvent",
-    "print",
-    "prompt",
-    "queueMicrotask",
-    "requestAnimationFrame",
-    "requestIdleCallback",
-    "ResizeObserver",
-    "screen",
-    "self",
-    "sessionStorage",
-    "setInterval",
-    "setTimeout",
-    "ShadowRoot",
-    "TouchEvent",
-    "WebGL2RenderingContext",
-    "WebGLRenderingContext",
-    "WebSocket",
-    "window",
-    "XMLHttpRequest",
-];
-
-/// Server-only globals (WinterCG extensions, ESM-based).
-static NODE_GLOBALS: &[&str] = &["Buffer", "clearImmediate", "process", "setImmediate"];
-
-/// Vue globals ($refs, $emit, $slots, $attrs, $el, etc.).
-static VUE_GLOBALS: &[&str] = &[
-    "$attrs",
-    "$data",
-    "$el",
-    "$emit",
-    "$forceUpdate",
-    "$nextTick",
-    "$options",
-    "$parent",
-    "$props",
-    "$refs",
-    "$root",
-    "$slots",
-    "$watch",
-];
-
-/// `ParamNames` (the owned form stored in each global scope) built once from the
-/// static name lists. Each `enter_*_scope` consumes an owned list, so we clone
-/// per file — but the `CompactString` values (all short enough to stay inline)
-/// are constructed only once here rather than per file.
-static BROWSER_GLOBAL_NAMES: LazyLock<ParamNames> =
-    LazyLock::new(|| build_global_names(BROWSER_GLOBALS));
-static NODE_GLOBAL_NAMES: LazyLock<ParamNames> = LazyLock::new(|| build_global_names(NODE_GLOBALS));
-static VUE_GLOBAL_NAMES: LazyLock<ParamNames> = LazyLock::new(|| build_global_names(VUE_GLOBALS));
-
-/// Materialize a `&'static [&str]` into the owned `ParamNames` form.
-fn build_global_names(names: &'static [&'static str]) -> ParamNames {
-    names
-        .iter()
-        .map(|&name| CompactString::const_new(name))
-        .collect()
-}
-
-/// Setup global scopes hierarchy:
-/// - ~universal (JS globals) - root, @0:0 (meta)
-/// - ~vue (Vue globals) - parent: ~universal, @0:0 (meta)
-/// - ~mod (module = SFC) - parent: ~universal, covers entire source
-fn setup_global_scopes(scopes: &mut ScopeChain, source_len: u32) {
-    // Root is already ~js (JsGlobalUniversal) with common globals
-    // Current scope is root (~js)
-
-    // !client - Browser-only globals (WHATWG Living Standard + HTML timers)
-    // Used as parent for onMounted, onUnmounted, etc.
-    scopes.enter_js_global_scope(
-        JsGlobalScopeData {
-            runtime: JsRuntime::Browser,
-            globals: BROWSER_GLOBAL_NAMES.clone(),
-        },
-        0,
-        0,
-    );
-    scopes.exit_scope(); // Back to ~univ
-
-    // #server - Server-only globals (WinterCG extensions, ESM-based)
-    // Reserved for future SSR/Server Components support
-    scopes.enter_js_global_scope(
-        JsGlobalScopeData {
-            runtime: JsRuntime::Node,
-            globals: NODE_GLOBAL_NAMES.clone(),
-        },
-        0,
-        0,
-    );
-    scopes.exit_scope(); // Back to ~univ
-
-    // ~vue - Vue globals (parent: ~univ, meta scope)
-    scopes.enter_vue_global_scope(
-        VueGlobalScopeData {
-            globals: VUE_GLOBAL_NAMES.clone(),
-        },
-        0,
-        0,
-    );
-    scopes.exit_scope(); // Back to ~univ
-
-    // ~mod - module scope (parent: ~js, covers entire SFC)
-    scopes.enter_module_scope(0, source_len);
-    // Stay in module scope - setup/plain will be created as children
-}
-
-/// Parse script setup source code using OXC parser with an optional generic parameter.
-///
-/// `generic` is the value from `<script setup generic="T">` attribute, if present.
-///
-/// This is a high-performance alternative to string-based analysis,
-/// providing accurate AST-based detection with proper span tracking.
-pub fn parse_script_setup_with_generic(source: &str, generic: Option<&str>) -> ScriptParseResult {
-    let allocator = Allocator::default();
-    let source_type = SourceType::from_path("script.ts").unwrap_or_default();
-
-    let ret = profile!(
-        "croquis.script_setup.oxc_parse",
-        Parser::new(&allocator, source, source_type).parse()
-    );
-
-    if ret.panicked {
-        return ScriptParseResult::default();
-    }
-
-    let source_len = source.len() as u32;
-
-    let mut result = ScriptParseResult {
-        bindings: BindingMetadata::script_setup(),
-        scopes: ScopeChain::with_capacity(16),
-        ..Default::default()
-    };
-
-    // Setup global scope hierarchy (universal → mod)
-    profile!(
-        "croquis.script_setup.global_scopes",
-        setup_global_scopes(&mut result.scopes, source_len)
-    );
-
-    // Enter script setup scope (parent: ~mod)
-    result.scopes.enter_script_setup_scope(
-        ScriptSetupScopeData {
-            is_ts: true,
-            is_async: false,
-            generic: generic.map(CompactString::new),
-        },
-        0,
-        source_len,
-    );
-
-    // Process all statements
-    profile!("croquis.script_setup.walk_statements", {
-        for stmt in ret.program.body.iter() {
-            process::process_statement(&mut result, stmt, source);
-        }
-    });
-
-    // After every binding is known, demote any `type` / `interface` that
-    // references a setup-scope value via `typeof` so the virtual TS keeps
-    // it inside `__setup` instead of hoisting it to module scope.
-    profile!(
-        "croquis.script_setup.resolve_type_hoisting",
-        result.resolve_type_export_hoisting()
-    );
-
-    result
-}
-
-/// Parse script setup source code using OXC parser.
-///
-/// This is a high-performance alternative to string-based analysis,
-/// providing accurate AST-based detection with proper span tracking.
-pub fn parse_script_setup(source: &str) -> ScriptParseResult {
-    parse_script_setup_with_generic(source, None)
-}
-
-/// Parse non-script-setup (Options API) source code using OXC parser.
-pub fn parse_script(source: &str) -> ScriptParseResult {
-    parse_script_with_options(source, ScriptParserOptions::default())
-}
-
-/// Parse non-script-setup source code using OXC parser with explicit options.
-pub fn parse_script_with_options(source: &str, options: ScriptParserOptions) -> ScriptParseResult {
-    let allocator = Allocator::default();
-    let source_type = SourceType::from_path("script.ts").unwrap_or_default();
-
-    let ret = profile!(
-        "croquis.script_plain.oxc_parse",
-        Parser::new(&allocator, source, source_type).parse()
-    );
-
-    if ret.panicked {
-        return ScriptParseResult::default();
-    }
-
-    let source_len = source.len() as u32;
-
-    let mut result = ScriptParseResult {
-        bindings: BindingMetadata::new(), // Not script setup
-        scopes: ScopeChain::with_capacity(16),
-        is_non_setup_script: true, // Mark as non-setup script for violation detection
-        ..Default::default()
-    };
-
-    // Setup global scope hierarchy (universal → mod)
-    profile!(
-        "croquis.script_plain.global_scopes",
-        setup_global_scopes(&mut result.scopes, source_len)
-    );
-
-    // Enter non-script-setup scope (parent: ~mod)
-    result.scopes.enter_non_script_setup_scope(
-        NonScriptSetupScopeData {
-            is_ts: true,
-            has_define_component: false,
-        },
-        0,
-        source_len,
-    );
-
-    process::collect_options_api_component_metadata(
-        &mut result,
-        &ret.program,
-        options.options_api,
-        options.legacy_vue2,
-    );
-
-    // Process all statements
-    profile!("croquis.script_plain.walk_statements", {
-        for stmt in ret.program.body.iter() {
-            process::process_statement(&mut result, stmt, source);
-        }
-    });
-
-    // Mirror the setup path so non-setup scripts also keep typeof-anchored
-    // types adjacent to their value bindings in any downstream emitters.
-    profile!(
-        "croquis.script_plain.resolve_type_hoisting",
-        result.resolve_type_export_hoisting()
-    );
-
-    result
-}
+pub(crate) use result::{ReactiveGetterContext, ReactiveValueOrigin, RuntimeObjectLiteral};
+pub use result::{ScriptParseResult, ScriptParserOptions};
 
 #[cfg(test)]
 mod tests {
